#9 Added basic HTML and loaded images
I have written (or in the case of the plot section copied) the layout for the page. I have loaded the images that are in the static/images folder on my laptop.

For the images I have used the picture element and have given them a standard width so that there won't be any layout shifts. To the images that appear later in the screen I have also put a lazy loading so the first content paint isn't heavy. I am however stumbling upon making some of these images appear as an avif or webp. If I cannot solve this in 2 hours starting from 14:30 I'll make a bug issue.

I would like to ask @joostf or @krijnhoetmer to take a look at this code and see what they think of this page thus far. (it's not yet near ready)