#9 Batman CSS
I loaded in the fonts that I will be using for the headings and for the paragraphs. I have tested them and it works. I will leave them off for now as I want to perfect the HTML first so that it will be functional first.

I made the colours in HSL, but I want to make them so that if you support oklch and p3 you'll be able to see the colours better. It isn't the focus in the coming week most likely, but it will be the focus once the HTML is 'finished' (because it can always be better.)