#11 Basic setup menu
Now as it stands there is a break between small screens at 600px so that the menu pops back up when it's at that point.

It is tab'able and it has a distinct focus on it. I have started on a hover animation and I will continue this animation futher and add other states as well as shown in #4